code/python/iothub_device/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
arrow==0.15.8             # via -r requirements.in
azure-iot-device==2.3.0   # via -r requirements.in
azure-iot-hub==2.2.3      # via -r requirements.in
certifi==2020.6.20        # via msrest, requests, uamqp
chardet==3.0.4            # via requests
deprecation==2.1.0        # via azure-iot-device
docopt==0.6.2             # via -r requirements.in
idna==2.10                # via requests
isodate==0.6.0            # via msrest
janus==0.4.0              # via azure-iot-device
msrest==0.6.19            # via azure-iot-hub
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.4           # via deprecation
paho-mqtt==1.5.1          # via azure-iot-device
pyparsing==2.4.7          # via packaging
pysocks==1.7.1            # via azure-iot-device
python-dateutil==2.8.1    # via arrow
requests-oauthlib==1.3.0  # via msrest
requests-unixsocket==0.2.0  # via azure-iot-device
requests==2.24.0          # via -r requirements.in, azure-iot-device, msrest, requests-oauthlib, requests-unixsocket
six==1.15.0               # via azure-iot-device, isodate, packaging, python-dateutil, uamqp
uamqp==1.2.12             # via azure-iot-hub
urllib3==1.25.10          # via azure-iot-device, requests, requests-unixsocket
